docker/API/SDXL/requirements.txt
accelerate==0.23.0
asgiref==3.5.2
brotlipy==0.7.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.0.4
cryptography==41.0.3
diffusers==0.21.3
Django==3.2.5
djangorestframework==3.14.0
filelock==3.9.0
fsspec==2023.9.2
gmpy2==2.1.2
huggingface-hub==0.17.3
idna==3.4
importlib-metadata==6.8.0
invisible-watermark==0.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
opencv-python==4.8.1.78
packaging==23.1
Pillow==9.4.0
pip==23.2.1
psutil==5.9.5
psycopg2==2.9.3
pycparser==2.21
pyOpenSSL==23.2.0
PySocks==1.7.1
pytz==2022.7
PyWavelets==1.4.1
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
safetensors==0.3.3
setuptools==68.0.0
sqlparse==0.4.3
sympy==1.11.1
tokenizers==0.13.3
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.66.1
transformers==4.33.3
triton==2.0.0
typing_extensions==4.7.1
urllib3==1.26.16
wheel==0.41.2
zipp==3.17.0
